Move AtomMapper into separate file; separate out tests.
